requirements.txt
alembic==0.6.0
Babel==1.3
beautifulsoup4==4.3.1
blinker==1.3
cx-Oracle==5.1.2
Flask-Admin==1.0.6
Flask-Principal==0.4.0
Flask-Script==0.6.2
Flask-SQLAlchemy==1.0
Flask-WTF==0.9.1
Flask==0.10.1
inflect==0.2.4
itsdangerous==0.23
Jinja2==2.7.1
Mako==0.9.0
MarkupSafe==0.18
mock==1.0.1
path.py==4.3
py==1.4.15
pytest==2.3.5
python-dateutil==2.1
pytz==2013d
six==1.4.1
sqlacodegen==1.1.1
SQLAlchemy==0.8.2
waitress==0.8.7
WebOb==1.2.3
WebTest==2.0.7
Werkzeug==0.9.4
WTForms==1.0.4
